+++ src/main/java/edu/ie3/datamodel/io/factory/input/participant/PvInputFactory.java
@@ -5,9 +5,9 @@
 */
 package edu.ie3.datamodel.io.factory.input.participant;
 
-import edu.ie3.datamodel.io.factory.input.NodeAssetInputEntityData;
 import edu.ie3.datamodel.models.OperationTime;
 import edu.ie3.datamodel.models.StandardUnits;
+import edu.ie3.datamodel.models.input.EmInput;
 import edu.ie3.datamodel.models.input.NodeInput;
 import edu.ie3.datamodel.models.input.OperatorInput;
 import edu.ie3.datamodel.models.input.system.PvInput;
@@ -19,7 +19,7 @@ import javax.measure.quantity.Power;
 import tech.units.indriya.ComparableQuantity;
 
 public class PvInputFactory
-    extends SystemParticipantInputEntityFactory<PvInput, NodeAssetInputEntityData> {
+    extends SystemParticipantInputEntityFactory<PvInput, SystemParticipantEntityData> {
   private static final String ALBEDO = "albedo";
   private static final String AZIMUTH = "azimuth";
   private static final String ETA_CONV = "etaConv";
@@ -43,13 +43,14 @@ public class PvInputFactory
 
   @Override
   protected PvInput buildModel(
-      NodeAssetInputEntityData data,
+      SystemParticipantEntityData data,
       UUID uuid,
       String id,
       NodeInput node,
       ReactivePowerCharacteristic qCharacteristics,
       OperatorInput operator,
       OperationTime operationTime) {
+    final EmInput em = data.getEm().orElse(null);
     final double albedo = data.getDouble(ALBEDO);
     final ComparableQuantity<Angle> azimuth = data.getQuantity(AZIMUTH, StandardUnits.AZIMUTH);
     final ComparableQuantity<Dimensionless> etaConv =
@@ -69,6 +70,7 @@ public class PvInputFactory
         operationTime,
         node,
         qCharacteristics,
+        em,
         albedo,
         azimuth,
         etaConv,