+++ src/main/java/edu/ie3/datamodel/models/input/container/SubGridContainer.java
@@ -24,10 +24,9 @@ public class SubGridContainer extends GridContainer {
       int subnet,
       RawGridElements rawGrid,
       SystemParticipants systemParticipants,
-      EnergyManagementUnits emUnits,
       GraphicElements graphics)
       throws InvalidGridException {
-    super(gridName, rawGrid, systemParticipants, emUnits, graphics);
+    super(gridName, rawGrid, systemParticipants, graphics);
     this.subnet = subnet;
     this.predominantVoltageLevel = ContainerUtils.determinePredominantVoltLvl(rawGrid, subnet);
   }
@@ -113,12 +112,7 @@ public class SubGridContainer extends GridContainer {
     @Override
     public SubGridContainer build() throws InvalidGridException {
       return new SubGridContainer(
-          getGridName(),
-          subnet,
-          getRawGrid(),
-          getSystemParticipants(),
-          getEmUnits(),
-          getGraphics());
+          getGridName(), subnet, getRawGrid(), getSystemParticipants(), getGraphics());
     }
   }
 }