+++ src/main/java/edu/ie3/datamodel/models/result/connector/ConnectorResult.java
@@ -88,6 +88,7 @@ public abstract class ConnectorResult extends ResultEntity {
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
+    if (!super.equals(o)) return false;
     ConnectorResult that = (ConnectorResult) o;
     return iAMag.equals(that.iAMag)
         && iAAng.equals(that.iAAng)